backend/requirements.txt
Django~=5.1.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
django-filter>=24.1
django-cors-headers>=4.3.1
django-environ>=0.11.2
django-countries>=7.6.0

psycopg>=3.1.8
requests>=2.32.3

Pillow>=11.0.0

Stripe>=11.3.0